removed deprecated data.py files from some packages
