--- pyworkflow/em/packages/grigoriefflab/data.py
@@ -1,37 +1,0 @@
-# **************************************************************************
-# *
-# * Authors:     Antonio Poza Ballesteros (apoza@cnb.csic.es)
-# *
-# * Unidad de  Bioinformatica of Centro Nacional de Biotecnologia , CSIC
-# *
-# * This program is free software; you can redistribute it and/or modify
-# * it under the terms of the GNU General Public License as published by
-# * the Free Software Foundation; either version 2 of the License, or
-# * (at your option) any later version.
-# *
-# * This program is distributed in the hope that it will be useful,
-# * but WITHOUT ANY WARRANTY; without even the implied warranty of
-# * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# * GNU General Public License for more details.
-# *
-# * You should have received a copy of the GNU General Public License
-# * along with this program; if not, write to the Free Software
-# * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
-# * 02111-1307  USA
-# *
-# *  All comments concerning this program package may be sent to the
-# *  e-mail address 'apoza@cnb.csic.es'
-# *
-# **************************************************************************
-"""
-This modules contains basic hierarchy
-for specific Xmipp3 EM data objects
-"""
-
-from pyworkflow.em import *   
-    
-class CTFFindModel(CTFModel):
-    """ CTFModel wrapper around the result file of ctffind3"""
-    pass
-        
-
